Don't generate assembly cache for CoreCLR
