+++ src/Xamarin.Android.Build.Tasks/Utilities/MarshalMethodsNativeAssemblyGenerator.cs
@@ -17,7 +17,7 @@ using CecilParameterDefinition = global::Mono.Cecil.ParameterDefinition;
 
 namespace Xamarin.Android.Tasks
 {
-	partial class MarshalMethodsNativeAssemblyGenerator : LlvmIrComposer
+	abstract partial class MarshalMethodsNativeAssemblyGenerator : LlvmIrComposer
 	{
 		const string GetFunctionPointerVariableName = "get_function_pointer";
 
@@ -159,7 +159,7 @@ namespace Xamarin.Android.Tasks
 			public string name;
 		}
 
-		sealed class AssemblyCacheState
+		protected sealed class AssemblyCacheState
 		{
 			public Dictionary<string, uint>? AsmNameToIndexData32;
 			public Dictionary<uint, (string name, uint index)> Hashes32;
@@ -229,7 +229,6 @@ namespace Xamarin.Android.Tasks
 		};
 
 		readonly ICollection<string> uniqueAssemblyNames;
-		readonly int numberOfAssembliesInApk;
 
 		StructureInfo marshalMethodsManagedClassStructureInfo;
 		StructureInfo marshalMethodNameStructureInfo;
@@ -246,11 +245,10 @@ namespace Xamarin.Android.Tasks
 		/// <summary>
 		/// Constructor to be used ONLY when marshal methods are DISABLED
 		/// </summary>
-		public MarshalMethodsNativeAssemblyGenerator (TaskLoggingHelper log, AndroidTargetArch targetArch, int numberOfAssembliesInApk, ICollection<string> uniqueAssemblyNames)
+		protected MarshalMethodsNativeAssemblyGenerator (TaskLoggingHelper log, AndroidTargetArch targetArch, ICollection<string> uniqueAssemblyNames)
 			: base (log)
 		{
 			this.targetArch = targetArch;
-			this.numberOfAssembliesInApk = numberOfAssembliesInApk;
 			this.uniqueAssemblyNames = uniqueAssemblyNames ?? throw new ArgumentNullException (nameof (uniqueAssemblyNames));
 			generateEmptyCode = true;
 			defaultCallMarker = LlvmIrCallMarker.Tail;
@@ -259,10 +257,9 @@ namespace Xamarin.Android.Tasks
 		/// <summary>
 		/// Constructor to be used ONLY when marshal methods are ENABLED
 		/// </summary>
-		public MarshalMethodsNativeAssemblyGenerator (TaskLoggingHelper log, int numberOfAssembliesInApk, ICollection<string> uniqueAssemblyNames, NativeCodeGenStateObject codeGenState, bool managedMarshalMethodsLookupEnabled)
+		protected MarshalMethodsNativeAssemblyGenerator (TaskLoggingHelper log, ICollection<string> uniqueAssemblyNames, NativeCodeGenStateObject codeGenState, bool managedMarshalMethodsLookupEnabled)
 			: base (log)
 		{
-			this.numberOfAssembliesInApk = numberOfAssembliesInApk;
 			this.uniqueAssemblyNames = uniqueAssemblyNames ?? throw new ArgumentNullException (nameof (uniqueAssemblyNames));
 			this.codeGenState = codeGenState ?? throw new ArgumentNullException (nameof (codeGenState));
 			this.managedMarshalMethodsLookupEnabled = managedMarshalMethodsLookupEnabled;
@@ -601,7 +598,8 @@ namespace Xamarin.Android.Tasks
 			MapStructures (module);
 
 			Init ();
-			AddAssemblyImageCache (module, out AssemblyCacheState acs);
+			AssemblyCacheState acs = CreateAssemblyCache ();
+			AddAssemblyImageCache (module, acs);
 
 			// class cache
 			module.AddGlobalVariable ("marshal_methods_number_of_classes", (uint)classes.Count, LlvmIrVariableOptions.GlobalConstant);
@@ -981,16 +979,9 @@ namespace Xamarin.Android.Tasks
 			}
 		}
 
-		// TODO: this should probably be moved to a separate writer, since not only marshal methods use the cache
-		void AddAssemblyImageCache (LlvmIrModule module, out AssemblyCacheState acs)
+		AssemblyCacheState CreateAssemblyCache ()
 		{
-			var assembly_image_cache = new LlvmIrGlobalVariable (typeof(List<IntPtr>), "assembly_image_cache", LlvmIrVariableOptions.GlobalWritable) {
-				ZeroInitializeArray = true,
-				ArrayItemCount = (ulong)numberOfAssembliesInApk,
-			};
-			module.Add (assembly_image_cache);
-
-			acs = new AssemblyCacheState {
+			var acs = new AssemblyCacheState {
 				AsmNameToIndexData32 = new Dictionary<string, uint> (StringComparer.Ordinal),
 				Indices32 = new List<uint> (),
 
@@ -1055,84 +1046,10 @@ namespace Xamarin.Android.Tasks
 				acs.AsmNameToIndexData64.Add (name, idx);
 			}
 
-			var assembly_image_cache_hashes = new LlvmIrGlobalVariable (typeof(List<ulong>), "assembly_image_cache_hashes", LlvmIrVariableOptions.GlobalConstant) {
-				Comment = " Each entry maps hash of an assembly name to an index into the `assembly_image_cache` array",
-				BeforeWriteCallback = UpdateAssemblyImageCacheHashes,
-				BeforeWriteCallbackCallerState = acs,
-				GetArrayItemCommentCallback = GetAssemblyImageCacheItemComment,
-				GetArrayItemCommentCallbackCallerState = acs,
-				NumberFormat = LlvmIrVariableNumberFormat.Hexadecimal,
-			};
-			module.Add (assembly_image_cache_hashes);
-
-			var assembly_image_cache_indices = new LlvmIrGlobalVariable (typeof(List<uint>), "assembly_image_cache_indices", LlvmIrVariableOptions.GlobalConstant) {
-				WriteOptions = LlvmIrVariableWriteOptions.ArrayWriteIndexComments | LlvmIrVariableWriteOptions.ArrayFormatInRows,
-				BeforeWriteCallback = UpdateAssemblyImageCacheIndices,
-				BeforeWriteCallbackCallerState = acs,
-			};
-			module.Add (assembly_image_cache_indices);
-		}
-
-		void UpdateAssemblyImageCacheHashes (LlvmIrVariable variable, LlvmIrModuleTarget target, object? callerState)
-		{
-			AssemblyCacheState acs = EnsureAssemblyCacheState (callerState);
-			object value;
-			Type type;
-
-			if (target.Is64Bit) {
-				value = acs.Keys64;
-				type = typeof(List<ulong>);
-			} else {
-				value = acs.Keys32;
-				type = typeof(List<uint>);
-			}
-
-			LlvmIrGlobalVariable gv = EnsureGlobalVariable (variable);
-			gv.OverrideTypeAndValue (type, value);
-		}
-
-		string? GetAssemblyImageCacheItemComment (LlvmIrVariable v, LlvmIrModuleTarget target, ulong index, object? value, object? callerState)
-		{
-			AssemblyCacheState acs = EnsureAssemblyCacheState (callerState);
-
-			string name;
-			uint i;
-			if (target.Is64Bit) {
-				var v64 = (ulong)value;
-				name = acs.Hashes64[v64].name;
-				i = acs.Hashes64[v64].index;
-			} else {
-				var v32 = (uint)value;
-				name = acs.Hashes32[v32].name;
-				i = acs.Hashes32[v32].index;
-			}
-
-			return $" {index}: {name} => {i}";
-		}
-
-		void UpdateAssemblyImageCacheIndices (LlvmIrVariable variable, LlvmIrModuleTarget target, object? callerState)
-		{
-			AssemblyCacheState acs = EnsureAssemblyCacheState (callerState);
-			object value;
-
-			if (target.Is64Bit) {
-				value = acs.Indices64;
-			} else {
-				value = acs.Indices32;
-			}
-
-			LlvmIrGlobalVariable gv = EnsureGlobalVariable (variable);
-			gv.OverrideTypeAndValue (variable.Type, value);
-		}
-
-		AssemblyCacheState EnsureAssemblyCacheState (object? callerState)
-		{
-			var acs = callerState as AssemblyCacheState;
-			if (acs == null) {
-				throw new InvalidOperationException ("Internal error: construction state expected but not found");
-			}
-
 			return acs;
 		}
+
+		protected virtual void AddAssemblyImageCache (LlvmIrModule module, AssemblyCacheState acs)
+		{}
 	}
 }
