+++ src/Xamarin.Android.Build.Tasks/Utilities/MarshalMethodsNativeAssemblyGeneratorCoreCLR.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+using Microsoft.Build.Utilities;
+using Xamarin.Android.Tools;
+
+namespace Xamarin.Android.Tasks;
+
+class MarshalMethodsNativeAssemblyGeneratorCoreCLR : MarshalMethodsNativeAssemblyGenerator
+{
+	/// <summary>
+	/// Constructor to be used ONLY when marshal methods are DISABLED
+	/// </summary>
+	public MarshalMethodsNativeAssemblyGeneratorCoreCLR (TaskLoggingHelper log, AndroidTargetArch targetArch, ICollection<string> uniqueAssemblyNames)
+		: base (log, targetArch, uniqueAssemblyNames)
+	{}
+
+	public MarshalMethodsNativeAssemblyGeneratorCoreCLR (TaskLoggingHelper log, ICollection<string> uniqueAssemblyNames, NativeCodeGenStateObject codeGenState, bool managedMarshalMethodsLookupEnabled)
+		: base (log, uniqueAssemblyNames, codeGenState, managedMarshalMethodsLookupEnabled)
+	{}
+}
