+++ src/native/clr/xamarin-app-stub/application_dso_stub.cc
@@ -170,27 +170,6 @@ DSOApkEntry dso_apk_entries[2] {};
 // Support for marshal methods
 //
 #if defined (RELEASE)
-void* assembly_image_cache[] = {
-	nullptr,
-	nullptr,
-
-};
-
-// Each element contains an index into `assembly_image_cache`
-const uint32_t assembly_image_cache_indices[] = {
-	0,
-	1,
-	1,
-	1,
-};
-
-// hashes point to indices in `assembly_image_cache_indices`
-const xamarin::android::hash_t assembly_image_cache_hashes[] = {
-	0,
-	1,
-	2,
-	3,
-};
 
 uint32_t marshal_methods_number_of_classes = 2;
 MarshalMethodsManagedClass marshal_methods_class_cache[] = {
